requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.10.5
charset-normalizer==3.4.3
datasets==4.1.1
dill==0.4.0
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
jiter==0.11.0
multidict==6.6.4
multiprocess==0.70.16
numpy==2.2.6
openai==2.1.0
packaging==25.0
pandas==2.3.3
propcache==0.4.0
pyarrow==21.0.0
pydantic==2.11.10
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
yarl==1.21.0
